Fix: GCC and clang compilation warnings
* Fixed long to double casts.
* Dereferenced the ac->isnull variable.
* Respect return statements.

Note: One warning still persists(clang) which is
propably fixed upstream LLVM.

+++ src/backend/utils/adt/agtype.c
@@ -2367,7 +2367,7 @@ static agtype_value* agtype_build_map_as_agtype_value(FunctionCallInfo fcinfo)
     nargs = extract_variadic_args(fcinfo, 0, true, &args, &types, &nulls);
 
     if (nargs < 0)
-        PG_RETURN_NULL();
+      PG_RETURN_NULL();
 
     if (nargs % 2 != 0)
     {
@@ -5215,7 +5215,7 @@ Datum age_tointeger(PG_FUNCTION_ARGS)
             float4 f = DatumGetFloat4(arg);
 
             if (isnan(f) || isinf(f) ||
-                f < PG_INT64_MIN || f > PG_INT64_MAX)
+                f < PG_INT64_MIN || f > (double)PG_INT64_MAX)
                 PG_RETURN_NULL();
 
             result = (int64) f;
@@ -5225,7 +5225,7 @@ Datum age_tointeger(PG_FUNCTION_ARGS)
             float8 f = DatumGetFloat8(arg);
 
             if (isnan(f) || isinf(f) ||
-                f < PG_INT64_MIN || f > PG_INT64_MAX)
+                f < PG_INT64_MIN || f > (double)PG_INT64_MAX)
                 PG_RETURN_NULL();
 
             result = (int64) f;
@@ -5238,7 +5238,7 @@ Datum age_tointeger(PG_FUNCTION_ARGS)
                 numeric_float8_no_overflow, arg));
 
             if (isnan(f) || isinf(f) ||
-                f < PG_INT64_MIN || f > PG_INT64_MAX)
+                f < PG_INT64_MIN || f > (double)PG_INT64_MAX)
                 PG_RETURN_NULL();
 
             result = (int64) f;
@@ -5270,7 +5270,7 @@ Datum age_tointeger(PG_FUNCTION_ARGS)
                  * return null.
                  */
                 if (!is_valid || isnan(f) || isinf(f) ||
-                    f < PG_INT64_MIN || f > PG_INT64_MAX)
+                    f < PG_INT64_MIN || f > (double)PG_INT64_MAX)
                     PG_RETURN_NULL();
 
                 result = (int64) f;
@@ -5302,7 +5302,7 @@ Datum age_tointeger(PG_FUNCTION_ARGS)
             float f = agtv_value->val.float_value;
 
             if (isnan(f) || isinf(f) ||
-                f < PG_INT64_MIN || f > PG_INT64_MAX)
+                f < PG_INT64_MIN || f > (double)PG_INT64_MAX)
                 PG_RETURN_NULL();
 
             result = (int64) f;
@@ -5316,7 +5316,7 @@ Datum age_tointeger(PG_FUNCTION_ARGS)
                 numeric_float8_no_overflow, num));
 
             if (isnan(f) || isinf(f) ||
-                f < PG_INT64_MIN || f > PG_INT64_MAX)
+                f < PG_INT64_MIN || f > (double)PG_INT64_MAX)
                 PG_RETURN_NULL();
 
             result = (int64) f;
@@ -5346,7 +5346,7 @@ Datum age_tointeger(PG_FUNCTION_ARGS)
                  * return null.
                  */
                 if (!is_valid || isnan(f) || isinf(f) ||
-                    f < PG_INT64_MIN || f > PG_INT64_MAX)
+                    f < PG_INT64_MIN || f > (double)PG_INT64_MAX)
                     PG_RETURN_NULL();
 
                 result = (int64) f;
